Changing short by int

+++ src/main/java/com/castaland/utils/Dices.java
@@ -2,12 +2,12 @@ package com.castaland.utils;
 
 public class Dices {
 
-    public static short throwDice() {
-        return (short) Primes.randInt(1, 6);
+    public static int throwDice() {
+        return Primes.randInt(1, 6);
     }
 
-    public static short[] throwDices(int num) {
-        short[] result = new short[num];
+    public static int[] throwDices(int num) {
+        int[] result = new int[num];
         for(int i = 0; i<num; i++) {
             result[i] = throwDice();
         }
